requirements.txt
anyio==3.6.1
bson==0.5.10
fastapi==0.78.0
idna==3.3
pydantic==1.9.1
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing-extensions==4.2.0
